B: Stop store emulation when API is disabled
When API was not enabled, the store emulation was started, but never
stopped. When loaded in the admin, this caused admin template files
trying to load from the frontend area, which resulted in the following error:
Warning: include(): Filename cannot be empty  in /Users/maikelkoek/Sites/fitforme.nl/app/code/core/Mage/Core/Block/Template.php on line 242

Also, when an exception occurs, we now stop the emulation in the try-catch.