Merged PR 5517: encrypt transactional passwords and token
Pointed to develop branch

encrypt transactional passwords and token
Now the transactional SMTP passwords can be viewed in admin and it’s value is encrypted into the database.
The upgrade script should take care of updating the existing values as the client should get the correct value.

Related work items: #57864, #64072