Merged PR 6747: Merge calin/catalog-visibility-fix to develop
This should fix the catalog visibility when individual is selecte from the setttings. Email catalog collection don't have any information about the product visibility so when the limit is set it will select first bulk of products even the products are not selected to be imported and it will bulk the limit.

*    this will exclude the fix for default option that should not be included when other options are selected
*    the filter for the collection should be on the catalog collection rathen on email_catalog, this will fix the missing imported products.

Related work items: #67520