Merged PR 12647: Skip product attributes if their keys are not valid
Skip product attributes if their keys are not valid.

The EC API has some restrictions on key names for insight data. During catalog sync, if we have invalid keys and attempt to sync single products, the import fails. This fix ensures that the product is synced, and any invalid keys are skipped when collecting the insight data payload.

To test a bundle product [id 447]:

- At L.149 insert `$options[24]->setDefaultTitle('Vælg dato');`

To test a configurable product [id 417]:

- At L.182 insert `$productAttributeOptions[0]['label'] = 'Vælg dato';`

Related work items: #85311