Merged PR 7005: Add to automation if object is new, or we've had a subscriber status change
Enrol subscriber to automation if the 'object is new' or if their subscriber status has changed to subscribed.

For Magento instances which are handling the opt in process, we were not enrolling the subscriber into the program when they become confirmed.

The code now checks if the object is new or if their subscription status has changed to subscribed.

Related work items: #68184