Merged PR 13597: Add index to order_id column in email_order table
Testing against a 3M-row `email_order` table does not support the suggestion that performance will be improved in the `sales_order_save_after` observer with the additional index. The collection query already uses the index on the `quote_id` column.

However, we may require a collection that only filters on `order_id`, in which case having the index will definitely improve performance over no index at all. With this change, we now have indexes on all columns in the `email_order` table.

Related work items: #88372