Merged PR 13731: Abandoned Carts now visible to limited permission users
From support ticket:
We often have clients limit our permissions in Magento to only items underneath the dotmailer extension. This is fine, but often means we can't access our own Abandoned Carts report - this happens in both M1 and M2. It's not unless they give us access to everything in Magento that we can see this report under Marketing Automation.

**Steps to Reproduce**
Create a permission template with only the available dotmailer permissions enabled. Assign the template to a user. Log in as that user and attempt to navigate to the abandoned cart report.

**Actual Result**
No abandoned cart report visible

**Expected Result**
Abandoned cart report visible

Related work items: #79713