Merged PR 13940: Take the right values per Website
At the moment although  when we syncing in Store Level we get The Default Values for all the products in EC.

**How to test**
Create a Website with a multiple views.
Then assign a product to this website and set different values (for this product) for each store view.

Make sure that the correct values are being syncing in EC

Related work items: #87937