Merged PR 14167: Url finder for product URLs
A new class to source correct product URLs for all types of products, in their correct store scope - for syncs and dynamic content links.

**To test catalog sync**

- Run catalog sync in default and store-level modes
- Check in particular the (not visible separately) children of configurable, grouped and bundle products
- Confirm store scope is applied in all cases
- Confirm URL rewrites are respected

**To test affected EDC links**

1. Abandoned cart basket
2. View wishlist
3. Any EDC that loads `template/connector/product` template files e.g. Wishlist related items
4. Reviews > Order contents. To test you need to set EC > Automation > Review reminder settings > Link to product page to 'Yes'.
5. One other for which you should definitely have data e.g. Recently viewed

To test the change to `price.phtml` you may need to comment out the conditions wrapping the code change. The change here relates to a link added to the 'as low as' text, which only shows in some scenarios.

Related work items: #89967