Merged PR 16286: Catalog items should use UrlFinder to get the image path
In M2 we built functionality to get the parent product image if the current product is a simple product, with no image specified. These changes should be applied to M1 during catalog sync.

**What Changed**
Functionality added to UrlFinder in order to grab the Parent Image of a simple product if this hasn't an image already.

**Why Changed**
During Catalog Sync if a simple product doesn't have an image then we get an invalid image url

**How to test**
- Checkout to this branch
- Try to sync a simple product that doesn't have an image , and make sure that we get its parent image .
- Make sure that we get the correct images for the rest of the products

Related work items: #97390