Merged PR 15557: Process product IDs in batches in catalog sync
The catalog sync is being changed so that it retrieves product IDs to be processed from the `email_catalog` table in advance of syncing by store. We are changing the notion of 'imported' products to 'processed' i.e. on each sync run we **process** 50 (if that is the batch size) products, of which we may only **sync** 5. By marking all 50 as processed, however, we ensure that we don't check them again on the next sync run.

## Why it's being changed

Some clients only sync a subset of products (e.g. configurable, visible products), and in these cases - especially with small sync batch sizes - it is possible for the sync to run out of qualifying products.

##What added
- new TS column `processed_at` in `email_catalog` table
- new backend-model with _beforeSave_ event to reset processed_at column if we change the `Include Products By Visibility` or `Include Products By Type` in Configuration section of our extension

##Why added
- Before this change we used to process non-imported products during catalog sync again and again.
- Now we will process the product only if never again checked/processed (that's why new column added).
- If we change the type of products we want to sync then `processed` field must be reset

## How to review / test this change

0) Upgrade
1) Configure the catalog sync to only sync a sub-set of products for this instance only grouped products (select all visibility cases)
2) Reset Catalog
3) First few times you run the catalog sync observe the `email_contact` and see that processed_at column is being filled and probably not the `imported`
4) Run sync a few times till you finally get an imported product (If all good first test passed :) )
5) Checkout to develop
6) Repeat 2,3,4
7) Normally the Catalog sync now is being stuck to first 50 products only and the rest never processed
(if yes 2nd test passed :) ) .
8) Checkout to this branch again and repeat 2,3,4
9) Go to EC > Configuration > Catalog Sync Settings and change "Include Products By Visibility"
Normally `processed` column should be reset for all non imported products
10) Repeat 9 but this time change "Include Products By Type"
(if  9 and 10 true, then 3rd test passed).

**Also**
- Test modifying an already-processed product and ensure that still works.
- Test both  Scope Store and Default Level

Related work items: #95202