Merged PR 16416: Remove markup attached to root.content.product-tooltip block
This PR removes some irrelevant markup from EDC blocks, which was being included as part of the core blank HTML template.

**To test**

Compare an EDC block with one on an environment without the change, and note that `<div>` and `<style>` tags which followed the opening `<body>` tag have been removed.

Related work items: #94137