Merged PR 16705: Fix dropIndex steps in 6.4.21 upgrade script
In our upgrade script for v6.4.21, we call `$installer->getTable` at the top of the script. Then, in each `dropIndex` step, we call it again, on the already-defined `$catalogTable`. This works fine unless the table has a prefix, in which case the table name passed to `dropIndex` becomes e.g. 'prefix_prefix_email_catalog'.

Related work items: #98857