Merged PR 16776: Fix URL redirect management on Enterprise
## What's being changed

A controller that handles the save action for URL rewrites.

## Why it's being changed

In M1 Enterprise edition [1.1.4], we have URL Redirect Management - whereas in regular M1 it is called URL Rewrite Management. There are additional differences between the core classes handling these sections of the admin:

- `Mage_Adminhtml_UrlrewriteController`
- `Enterprise_UrlRewrite_Adminhtml_UrlrewriteController`

The result is that URL Redirect Management in Enterprise is effectively being broken by our connector. This PR fixes it, enabling both to function depending on the platform.

## How to review / test this change

- In M1 Enterprise go to Catalog > URL Redirects
- Create, edit and save redirects
- Check that products are set unprocessed after save
- Repeat for regular M1

Related work items: #99145