Merged PR 16744: Mark ordered products as modified
Some clients noticed that they get wrong stock values in catalog sync.

This is happening because after the completion of a product purchase we don't
mark it again as `un-processed`. This means that every time a user buys a product
then we don't resync the new decreased stock value.

**What changed**
Now during order sync we mark ordered products as un-processed. By doing this we resync these products during catalog sync, so the new stock value will be updated.

**How to test**
- Simply complete a purchase of a product that is already imported into EC.
- After you run the order sync, then the product should be marked as processed = 0 in `email_catalog`

Related work items: #99017