Merged PR 23105: Protect order sync from bad SKUs
## What's being changed

We're try/catching when we load a product from the ProductRepository by SKU at the end of the order sync.

## Why it's being changed

It's possible that products in historical orders have SKUs that have now been changed in the catalog. When we come to mark products as unprocessed at the end of the order sync, we load them by SKU because at this point we're using the product data that is about to be sent to EC as insight data. We loaded the product earlier by its ID - and that worked - but loading it by its SKU doesn't.

Currently this throws an sql error:
```
SQLSTATE[42000]: Syntax error or access violation: 1064 You have an error in your SQL syntax
```
that derails the sync.

## How to review / test this change

- Place an order
- Go into sales_order_item and change the sku of one of the items you purchased to something that doesn't exist in the catalog (alternatively, change the SKU in the catalog to create the mismatch)
- Run order sync
- Check the log messages

Related work items: #120839